+++ src/main/java/org/truffleruby/core/CoreLibrary.java
@@ -453,6 +453,7 @@ public class CoreLibrary {
                 "Converter");
         final RubyModule truffleRubyModule = defineModule("TruffleRuby");
         defineClass(truffleRubyModule, objectClass, "AtomicReference");
+        defineClass(truffleRubyModule, objectClass, "ConcurrentMap");
         truffleModule = defineModule("Truffle");
         truffleInternalModule = defineModule(truffleModule, "Internal");
         graalErrorClass = defineClass(truffleModule, exceptionClass, "GraalError");
@@ -976,6 +977,7 @@ public class CoreLibrary {
             "/core/kernel.rb",
             "/core/lazy_rubygems.rb",
             "/core/truffle/boot.rb",
+            "/core/truffle/concurrent_map.rb",
             "/core/truffle/debug.rb",
             "/core/truffle/diggable.rb",
             "/core/truffle/encoding_operations.rb",